+++ src/super/i-static-page/modules/ss-helpers/page.spec.js
@@ -92,14 +92,14 @@ if ('std_style' in PATH) {
 \t
 \tel.setAttribute('rel', 'preload');
 \tel.setAttribute('as', 'style');
-\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['std_style']); })()));
+\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['std_style']); })()));
 \tdocument.head.appendChild(el);
 })();
 
 (function () {
 \tvar el = document.createElement('link');
 \t
-\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['std_style']); })()));
+\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['std_style']); })()));
 \tel.setAttribute('rel', 'stylesheet');
 \tel.setAttribute('media', 'print');
 \tel.setAttribute('onload', 'this.media=\\'all\\'; this.onload=null;');
@@ -124,14 +124,14 @@ function $__RENDER_ROOT() {
 \t
 \tel.setAttribute('rel', 'preload');
 \tel.setAttribute('as', 'style');
-\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['p-v4-components-demo_style']); })()));
+\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['p-v4-components-demo_style']); })()));
 \tdocument.head.appendChild(el);
 })();
 
 (function () {
 \tvar el = document.createElement('link');
 \t
-\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['p-v4-components-demo_style']); })()));
+\tel.setAttribute('href', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['p-v4-components-demo_style']); })()));
 \tel.setAttribute('rel', 'stylesheet');
 \tel.setAttribute('media', 'print');
 \tel.setAttribute('onload', 'this.media=\\'all\\'; this.onload=null;');
@@ -144,7 +144,7 @@ if ('std' in PATH) {
 (function () {
 \tvar el = document.createElement('script');
 \tel.async = false;
-\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['std']); })()));
+\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['std']); })()));
 \tdocument.head.appendChild(el);
 })();
 
@@ -176,7 +176,7 @@ if ('index-core' in PATH) {
 (function () {
 \tvar el = document.createElement('script');
 \tel.async = false;
-\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['index-core']); })()));
+\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['index-core']); })()));
 \tdocument.head.appendChild(el);
 })();
 
@@ -186,7 +186,7 @@ if ('vendor' in PATH) {
 (function () {
 \tvar el = document.createElement('script');
 \tel.async = false;
-\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['vendor']); })()));
+\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['vendor']); })()));
 \tdocument.head.appendChild(el);
 })();
 
@@ -195,7 +195,7 @@ if ('vendor' in PATH) {
 (function () {
 \tvar el = document.createElement('script');
 \tel.async = false;
-\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['p-v4-components-demo_tpl']); })()));
+\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['p-v4-components-demo_tpl']); })()));
 \tdocument.head.appendChild(el);
 })();
 
@@ -203,7 +203,7 @@ if ('vendor' in PATH) {
 (function () {
 \tvar el = document.createElement('script');
 \tel.async = false;
-\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['p-v4-components-demo']); })()));
+\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['p-v4-components-demo']); })()));
 \tdocument.head.appendChild(el);
 })();
 
@@ -212,7 +212,7 @@ if ('vendor' in PATH) {
 (function () {
 \tvar el = document.createElement('script');
 \tel.async = false;
-\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); } return concatURLs('/dist/client', PATH['webpack.runtime']); })()));
+\tel.setAttribute('src', ((function () { function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; } return concatURLs('/dist/client', PATH['webpack.runtime']); })()));
 \tdocument.head.appendChild(el);
 })();