+++ src/super/i-block/modules/activation/index.ts
@@ -12,6 +12,7 @@
  */
 
 import symbolGenerator from 'core/symbol';
+import { unwrap } from 'core/object/watch';
 import { runHook, callMethodFromComponent } from 'core/component';
 
 import type iBlock from 'super/i-block/i-block';
@@ -68,7 +69,10 @@ export function activate(component: iBlock, force?: boolean): void {
 				async function handler(route: typeof r.route, type: string): Promise<void> {
 					try {
 						if (type === 'hard') {
-							if (route !== r.route) {
+							const
+								actualRoute = unwrap(r.route) ?? r.route;
+
+							if (route !== actualRoute) {
 								await unsafe.promisifyOnce('setRoute', {
 									label: $$.activateAfterTransition
 								});